Update events to use actual provided data
Before data update events where triggered, whenever the data type was sent and the actual data item existed in the data storage. Which means that only updating a current measurement also triggered an update event for SoC if that was present in the store from an earlier update.

This is now changed and only considers the data elements that are actually provided in the payload.

Note: This does NOT consider if a datapoint has been deleted and not updated!